Move extensionPoint into common
