+++ src/vs/workbench/contrib/customEditor/common/extensionPoint.ts
